chore: clean test and add media manager as compsoer dependency
