+++ composer.json
@@ -31,6 +31,7 @@
     "oat-sa/extension-tao-delivery": ">=14.10.1",
     "oat-sa/extension-tao-outcomeui": ">=10.4.0",
     "oat-sa/extension-tao-test": ">=15.15.1",
+    "oat-sa/extension-tao-mediamanager": ">=12.32.1",
     "oat-sa/lib-tao-elasticsearch": ">=2.5.2"
   },
   "autoload": {
